(workspace-yaml-6) Add autodiscovery for files and modules
Summary:
This adds the code that takes a file or a module and finds
sole repository or pipeline symbol in it and if not, throws.

It was suggested that we also have a mode where we can just package
up all the top level definitions in a file and make a repository.

I decided not to do this for two reasons:

1) We would have to make a change to the core CodePointer abstract
to be able to point to entire module as opposed to a specific
symbol within that module

2) I'm not sure if its a good idea in essence. Seems like it might
be sensible to force repository creation when you have more than
one def

Depends on D3249

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3250