(bootstrap-autodiscovery-1) Bootstrap autodiscovery
Summary:
When booting up dagit or other host processes and
discovering repositories in other virtual environments, we
need to be able to iterator over a python module and
find repositories *in the user process*. This happens
before repository location handle creation. So we
add this api method do exactly that. Used in follow-on
diff to enable loading user processes from other
python virtual environments in a follow on diff.

Test Plan: BK

Reviewers: alangenfeld, max, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3326