[legacy-partitions-2] Change partition config picker to call external api upon selection replace dauphin with repo handle
Summary: We previously were loading the yaml / tags for every single partition, regardless of whether it was selected or not.  For partition sets with a lot of partitions, this became ungodly slow, because we're making an out-of-process call for every single one of these.  Instead, we're changing the picker so that it fetches the tags and run_config upon selection.  This requires a separate graphql query and different interaction UI to account for the increased latency.

Test Plan: Selected partition from partition selector, saw loading indicator, tags and config populated

Reviewers: alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3359