(workspace-cli-3) Make dagster pipeline list work with workspace.yaml format
Summary:
This makes dagster pipeline list work with the new workspace
format. This still relies on the dagster cli residing in the same
python environment as the user repository, either by being an in-process
location or an python environment location that happens to use
the same executable.

Depends on D3356

Test Plan: BK

Reviewers: alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3357