(workspace-cli-9) Port backfilll to workspace loading
Summary:
This is the final dagster cli command to port to new workspace
loading.

There is a slight behavior change here, which eliminates the feature
(unique to backfill) that prompts the user to select a pipeline if
many are available. Instead this uses the same resolution rules
as the rest of the commands. If we want to readd prompting, we should
do so in a way that cuts across all the commands.

This also simplified the test suite, which was doing some odd string
munging and transforms to go back and forth between different casing
and data structures for cli args.

Depends on D3389

Test Plan: BK

Reviewers: alangenfeld, sashank, prha

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3390