2/n Base schedule views off of schedule definition query
Summary:
This changes the schedule views to be based off of schedule definitions as the root, instead of schedule state. This is because schedule state is not the source of truth anymore, since we don't automatically run reconcile on dagit init.

On a `DauphinScheduleDefinition`, we can query for `DauphinScheduleState`, which was previously `RunnningSchedule`. This field is nullable - if it is null that means the schedule has not been reconciled yet.

Upcoming:
- Show schedule states in db for which there are no matching schedule definitions.
- Reconcile button

Test Plan: unit

Reviewers: alangenfeld, prha

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3411