Refactoring run creation pathway for better debugging and constraints
Summary:
We are seeing a snapshotting bug with some users in prod and
this is refactor that seemed reasonably while beginning to understand
the issue.

A couple changes

1) Renames snapshot_id to pipeline_snapshot_id in
sync_get_external_execution_plan to make it clear what type of
snapshot id is in play here

2) Extract a couple methods out of _construct_run_with_snapshots. The
persistence of the pipeline snapshot and the persistence of the
execution plan snapshot.

3) Eliminate an invalid state of _construct_run_with_snapshots where
the pipeline snapshot is None but the execution plan is not. This should
never occur.  The combination was broken anyways and would have
fataled because of a unbound local variable.

4) Added some more context to invariant violation being encountered in
prod.

5) Eliminate with_pipeline_snapshot_id and
with_execution_plan_snapshot_id on PipelineRun, which were only used
here.

Test Plan: BK

Reviewers: sashank, nate, prha

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D3574