Pin protobuf and grpcio-tools
Summary:
Unfortunately protobuf code generated by a particular version of grpcio-tools is not guaranteed to work with the range of protobuf versions required by that version of grpcio-tools. This diff pins grpcio-tools and protobuf to known compatible versions. The downside is that users who want to install dagster into an environment that has an existing grpc install will hit compatibility issues.

Resolves https://github.com/dagster-io/dagster/issues/2660

Test Plan: Manual

Reviewers: alangenfeld, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3661