Pipeline-scoped runs tab, cleaner cursor-paginated query utils
Summary:
This diff adds a new "Runs" tab to the pipeline-centric top navigation. It's similar to the instance-level Runs tab, but is quietly scoped to just the current pipeline.

Pipeline navigation is also present when you drill into a run. I think that's preferable since it lets you jump more easily to the playground or to explore the DAG. If the pipeline has been deleted you can still drill in, but moving off the Run to another one of the tabs gives you the "pipeline not found" state.

This diff also refactors the "cursor-based pagination" we use for the run history into a re-usable React hook and pairs it with a reusable pagination component for rendering the next / prev buttons.

{F178861}
{F178863}

Test Plan: Updated snapshots

Reviewers: alangenfeld, prha, schrockn

Reviewed By: prha, schrockn

Differential Revision: https://dagster.phacility.com/D3671