Composition pattern for GeneratorExit shenanigans
Summary:
I saw this again in one of @max's diffs and it still made me sad.

Again this is one of those where I'm not sure it's that much better but
at least this way the complexity is composed just a little more. Curious
what you two think.

Test Plan: BK

Reviewers: max, prha

Reviewed By: max, prha

Subscribers: max

Differential Revision: https://dagster.phacility.com/D3787