Remove ability to load modules resolved by the working directory
Summary:
By default, the current working directory is used to resolve modules.

This means that modules that are resolved when running dagit from one directory
might not resolve when resolved from another, even when using the same repo
args.  This is a problem with the current implementation of the cron scheduler,
which will not have the working directory context to resolve modules.

This diff removes that ability to make things more consistent, and adds a more
descriptive error message.

Test Plan: BK

Reviewers: schrockn, alangenfeld, sashank

Reviewed By: schrockn, alangenfeld

Differential Revision: https://dagster.phacility.com/D3670