Unit test buildkite image
Summary:
This adds a new image to our suite of images, the unit test
image. This preinstalls a number of common modules (esp the pandas/numpy
stack) that are commonly used across our unit tests and our test image.

The short term of goal of this is to speed builds. This brings the test
image build time (which is in the critical path) down from 2 minutes and
30-40 seconds to around 1 min, around a 90 seconds savings.

This should also serve as the base image for the unit test builds once
they in a separate BK queue. This should be a significant disk and
speed savings as these images are around 593MB rather than the >6GB
integration images.

This process is a bit manual right now and could be improved. However
it is working and I need to move on to other things, so I would rather
leave further work on this to another. (It would be a good intro issue
to learn the test pipeline)

Here is the relevant text in the README:

```

Similarly we have a similar system for the unit test images. These are more
modest and require less build times so we have not pushed the process
to a pipeline in buildkite.

1. Update the git has in ./buildkite/images/unit_snapshot_builder/Dockerfile
2. Run ./buildkite/images/build_all_unit_snapshots.py. This updates snapshot
   reqs files in ./buildkite
3. Manually update UNIT_IMAGE_VERSION.
   Run `python -c "import datetime; print(datetime.datetime.utcnow().strftime('%Y-%m-%dT%H%M%S'))"`
   and copy and paste output
4. Then run build_all_unit_images.py
5. Then run push_all_unit_images.py
6. Next you have to update the Dockerfile in dagster-test manually with value in UNIT_IMAGE_VERSION in the `FROM` directive.
7. Push the diff up and ensure it passes
```

Test Plan: BK

Reviewers: nate, catherinewu, dgibson, alangenfeld, max

Reviewed By: max

Subscribers: sashank

Differential Revision: https://dagster.phacility.com/D3802