Enable multi-selection (+cancel +deletion) of runs in the runs list #2661
Summary:
This diff adds checkboxes to the leftmost column of the runs list, and a checkbox at the top for select-all / unselect-all. A new actions menu beside the button allows you to quickly Cancel / Delete runs.

Right now this just makes individual graphQL queries in a for loop, and you get one "toast" for each result. This seems to work pretty well, but if you delete 30 runs you get a wall of toasts which makes you feel very productive but is not terribly pretty.

There's a bit of complexity to this because I wanted to allow the user to check some runs, click "Next Page" and check some more runs. This requires keeping the full Run objects in selection state rather than just run IDs, but I think it's fairly worth it. It means there are edge cases where we might try to submit a "terminate" mutation for a run that has already terminated, but whenever you're working with runs on the current page we use the latest data from that polling query, and out-of-date states could have happened anyway.

{F202058}
{F202059}

(In the second screenshot, if your selection includes some terminated and some cancelable runs, the delete / cancel commands will offer to perform the action on different numbers of runs)

Test Plan: Run tests

Reviewers: alangenfeld, prha, schrockn

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D3889