Remove nightly and deploy pipelines
Summary:
Now that we are in a good regular cadence of releasing on a weekly basis, the nightly builds don't feel like something we should continue to maintain.

This diff removes the buildkite pipeline; will follow-up to remove the nightly-specific code from our PyPI automation and version files.

Test Plan: ¯\_(ツ)_/¯

Reviewers: max, schrockn

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D3910