add linter for try/finally
Summary:
This is not sufficient for avoiding GC-related runtime errors caused
by yielding in finally blocks (still need to add logic to avoid yielding after
GeneratorExit), but it's a good opportunity to catch folks who have no idea.

Not sure if this is the right location for it - suggestions welcome.

Test Plan: created unit test, saw squigglies in vscode

Reviewers: schrockn, max

Reviewed By: max

Subscribers: sandyryza, alangenfeld

Differential Revision: https://dagster.phacility.com/D3893