Make ephemeral_grpc_server a class so that we can store it in the repository location
Summary: The plan is to instantiate these processes in the startup flow, put them on the repository location, and have them automatically clean themselves up during destruction (but also allow them to be used as context managers). Moving from a contextmanager function to a class with __enter__ and __exit__ (and __del__) implemented allows us to accomplish both of these goals.

Test Plan: BK

Reviewers: max, sashank, alangenfeld

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D3959