Return correct repository information when GRPC server is initialized with an attribute
Summary:
Before, if you pass in an attribute into the GRPC server as the target origin, it would assume that that attribute was also the repository name and incorrectly return it as the key in the list of LoadableRepositorySymbols it returned. Now, it pulls the actual name from the python environment like it does in the non-attribute path. I don't think this was causing any user-visible issues, but we should still return the right thing from list_repositories.

(The corresponding API in the CLI does not allow you to pass in an attribute, so this wasn't an issue there).

Test Plan: New BK test, was failing before in the GRPC path. Note that the grpc and non-grpc test cases are different because the non-GRPC path (harmlessly) uses the attribute as the key in the code pointer dict rather than the repository name

Reviewers: schrockn, alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3989